backend/requirements.txt
fastapi
uvicorn[standard]
motor
google-genai
pydantic
python-dotenv
duckdb
